backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
django-silk
djangorestframework-simplejwt
drf-spectacular
django-filter
django-cors-headers